refactor(archive): colocate export code in base/archive package
this'll make it easier to refactor & test in the future, while providing more
scope for what import & export should do: work with high-fidelity archives.
